+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Added optional placeholder prop to DayPicker
+- Select options for month and year in date and day picker
 
 # [2.0.0-beta.62] - 26-05-2020